+++ muna/beta/remote/__init__.py
@@ -1,8 +1,0 @@
-# 
-#   Muna
-#   Copyright © 2026 NatML Inc. All Rights Reserved.
-#
-
-from .endpoint import get_prediction_request, prediction_endpoint
-from .prediction import PredictionService
-from .schema import RemoteAcceleration, RemotePrediction, RemoteValue