+++ muna/types/__init__.py
@@ -4,7 +4,7 @@
 #
 
 from .parameter import BatchConfig, EnumerationMember, Parameter
-from .prediction import Acceleration, Prediction, PredictionResource
+from .prediction import Acceleration, Prediction, PredictionResource, RemotePrediction
 from .predictor import Predictor, PredictorAccess, PredictorStatus, Signature
 from .user import User
-from .value import Dtype, Value
+from .value import Dtype, RemoteValue, Value